+++ packages/builder/src/types.ts
@@ -1,9 +1,5 @@
-import {
-  AllowAnyTypeInformation,
-  PrimitiveType,
-  TypeInformation,
-} from "@rster/types";
 import { AllowVoidIfUndefined, NoUndefined } from "@rster/util";
+import { AllowAnyTypeInformation, PrimitiveType } from "@rster/types";
 import { RsterApiMethod } from "./method";
 import { RsterApiModule } from "./module";
 
@@ -13,7 +9,7 @@ export type AnyRsterApiMethod<key extends string = string> = RsterApiMethod<
 >;
 
 export type AnyParameterDeclaration = ParameterDeclaration<
-  TypeInformation<unknown>,
+  AllowAnyTypeInformation,
   ParameterList,
   ParameterList,
   ParameterList
@@ -25,8 +21,8 @@ export type ActionFunction<D extends AnyParameterDeclaration> = (
 
 export type ParameterList = {
   [key: string]:
-    | { type: TypeInformation<unknown>; required: true }
-    | { type: TypeInformation<unknown>; required: false };
+    | { type: AllowAnyTypeInformation; required: true }
+    | { type: AllowAnyTypeInformation; required: false };
 };
 
 /**
@@ -43,7 +39,7 @@ export type ParameterList = {
  * @param returns - The return type.
  */
 export interface ParameterDeclaration<
-  RETURNS extends TypeInformation<unknown>,
+  RETURNS extends AllowAnyTypeInformation,
   EXPECT_BODY extends ParameterList,
   EXPECT_QUERY extends ParameterList,
   EXPECT_PARAMS extends ParameterList