+++ packages/database-adapter-mongodb/changelog.md
@@ -1,5 +1,15 @@
 # Changelog for @rster/database-adapter-mongodb
 
+## 0.1.2
+
+### Patch Changes
+
+- Update util dependency
+- Updated dependencies [98fa3d0]
+- Updated dependencies
+  - @rster/types@0.1.8
+  - @rster/database@0.1.1
+
 ## v0.1.1
 
 - Add package index file