+++ packages/info/changelog.md
@@ -1,5 +1,24 @@
 # Changelog for @rster/info
 
+## 0.1.7
+
+### Patch Changes
+
+- Update util dependency
+- Updated dependencies [98fa3d0]
+- Updated dependencies
+  - @rster/types@0.1.8
+  - @rster/common@0.1.3
+  - @rster/basic@0.1.6
+
+## 0.1.6
+
+### Patch Changes
+
+- some small type-related patches
+- Updated dependencies
+  - @rster/types@0.1.6
+
 ## v0.1.5
 
 - Move declaration `optional` to `required` to make it consistent with `@rster/types`'s `object` type