+++ packages/proxy/changelog.md
@@ -1,5 +1,20 @@
 # Changelog for @rster/proxy
 
+## 0.1.1
+
+### Patch Changes
+
+- Update util dependency
+- Updated dependencies [98fa3d0]
+- Updated dependencies [3926989]
+- Updated dependencies
+  - @rster/types@0.1.8
+  - @rster/util@0.1.1
+  - @rster/builder@0.1.7
+  - @rster/common@0.1.3
+  - @rster/basic@0.1.6
+  - @rster/info@0.1.7
+
 ## v0.1.0
 
 - Initial release