+++ packages/util/changelog.md
@@ -1,5 +1,11 @@
 # Changelog for @rster/common
 
+## 0.1.1
+
+### Patch Changes
+
+- 3926989: Move Expand, NoUndefined, Extends types into util package
+
 ## v0.1.0
 
 - Initial release